// src/app/interfaces/cart.ts
// response of display is status and arr of books (cartitems)
import { Book } from './book';

export interface CartItem { // data in each book
    book: Book;
    quantity: number;
    _id: string; // cart item id generated by backend
}

export interface CartData {
    _id: string; // cart id generated by backend
    user: string; // user id
    items: CartItem[]; // arr of cart items
}

export interface CartResponse { // res(display,add,update,remove)
    status: string;
    data: CartData;
}

export interface CartCountResponse { // res of count
    status: string;
    data: {
        count: number;
    }
}